ref(mdc): move dataset-specific classes required by entities out of dataset sub-modules
